Add a C API `nwind_backtrace` around the local unwind API
This allows other projects not written in Rust to consume this
unwinding library. The API is pretty basic and only offers the
fundamental building blocks required for an external consumer
to drive the unwinding. There is no fancy lazy global static
address space, nor any thread local unwind context - configuring
that is left to the consumer of this API.